Move logger to utils
