+++ src/db/access-parser/run.ts
@@ -4,7 +4,7 @@ import ADODB from 'node-adodb';
 import { Connection, createConnection, getRepository, Repository } from 'typeorm';
 import config from '../prepare-db-config';
 import { EURINGs } from './EURING-access-tables';
-import { logger } from '../../configs/logger';
+import { logger } from '../../utils/logger';
 
 const accessConnection = ADODB.open(
   'Provider=Microsoft.ACE.OLEDB.12.0;Data Source=MBCRB(2007-30.01.2012).mdb;Persist Security Info=False;',
